requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.16.3
psycopg2-binary==2.8.2
pytz==2019.1
scipy==1.2.1
six==1.12.0
Werkzeug==0.15.4
whitenoise==3.2.2
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.3
psycopg2-binary==2.8.2
pytz==2019.1
scipy==1.2.1
six==1.12.0
sklearn==0.0
Werkzeug==0.15.4
whitenoise==3.2.2
aniso8601==6.0.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
MarkupSafe==1.1.1
numpy==1.16.3
psycopg2-binary==2.8.2
pytz==2019.1
PyYAML==5.1
scipy==1.2.1
six==1.12.0
Werkzeug==0.15.4
whitenoise==3.2.2
absl-py==0.7.1
aniso8601==6.0.0
astor==0.8.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gast==0.2.2
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
numpy==1.16.3
Pillow==6.0.0
protobuf==3.8.0
psycopg2-binary==2.8.2
pytz==2019.1
PyYAML==5.1
scikit-learn==0.21.1
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorflow==1.13.1
termcolor==1.1.0
Werkzeug==0.15.4
whitenoise==3.2.2
